fix(ci): python-coverage-shards — split 4 files under the 1500-line maintainability gate
PR #421 lease-fencing work pushed four first-party files past the
MAX_FIRST_PARTY_FILE_LINES=1500 guardrail enforced by
test_core_decomposition_maintainability.py, failing the coverage shard:

- src/awf/db/repositories/workspace_repo.py (1513): extract the
  monitor/execution claim+lease methods into workspace_repo_claims.py as
  module-level helpers, delegated from thin wrappers (mirrors the existing
  workspace_repo_host_ports.py split).
- src/awf/node/provisioner.py (1630): extract the two host-port admission
  checks into provisioner_host_ports_check.py, wired back via
  ProvisionerHostPortCheckMixin so call sites and super() overrides are
  unchanged (mirrors the executor *DelegatesMixin pattern).
- tests/.../test_repository_coverage_part_001.py (1607): move the four
  execution-claim-epoch tests into part_002.
- tests/unit/control/test_worker_scheduler_admission.py (1703): move the
  shared fakes/DB helpers into _worker_admission_support.py.

Behavior is unchanged; this is a mechanical decomposition.
